Refurbished SLg testsuite and fixed an error in createLRR method. Issues #37 and #38, PR #111
