+++ testsuite/tests/src/test/java/org/mobicents/diameter/stack/functional/slg/AbstractSLgClient.java
@@ -21,9 +21,6 @@
 
 package org.mobicents.diameter.stack.functional.slg;
 
-import java.io.InputStream;
-import java.util.concurrent.TimeUnit;
-
 import org.jdiameter.api.ApplicationId;
 import org.jdiameter.api.Avp;
 import org.jdiameter.api.AvpSet;
@@ -31,6 +28,7 @@ import org.jdiameter.api.IllegalDiameterStateException;
 import org.jdiameter.api.InternalException;
 import org.jdiameter.api.Mode;
 import org.jdiameter.api.OverloadException;
+import org.jdiameter.api.Request;
 import org.jdiameter.api.RouteException;
 import org.jdiameter.api.app.AppAnswerEvent;
 import org.jdiameter.api.app.AppRequestEvent;
@@ -38,26 +36,31 @@ import org.jdiameter.api.app.AppSession;
 import org.jdiameter.api.slg.ClientSLgSession;
 import org.jdiameter.api.slg.ClientSLgSessionListener;
 import org.jdiameter.api.slg.ServerSLgSession;
-import org.jdiameter.api.slg.events.ProvideLocationRequest;
-import org.jdiameter.api.slg.events.ProvideLocationAnswer;
-import org.jdiameter.api.slg.events.LocationReportRequest;
 import org.jdiameter.api.slg.events.LocationReportAnswer;
+import org.jdiameter.api.slg.events.LocationReportRequest;
+import org.jdiameter.api.slg.events.ProvideLocationAnswer;
+import org.jdiameter.api.slg.events.ProvideLocationRequest;
 import org.jdiameter.client.api.ISessionFactory;
+import org.jdiameter.common.impl.app.slg.LocationReportAnswerImpl;
 import org.jdiameter.common.impl.app.slg.ProvideLocationRequestImpl;
 import org.jdiameter.common.impl.app.slg.SLgSessionFactoryImpl;
 import org.mobicents.diameter.stack.functional.TBase;
+
+import java.io.InputStream;
+import java.util.concurrent.TimeUnit;
 
 /**
  *
  *@author <a href="mailto:fernando.mendioroz@gmail.com"> Fernando Mendioroz </a>
  *
  */
-public abstract class AbstractImmediateClient extends TBase implements ClientSLgSessionListener {
+public abstract class AbstractSLgClient extends TBase implements ClientSLgSessionListener {
 
   // NOTE: implementing NetworkReqListener since its required for stack to
   // know we support it... ech.
 
   protected ClientSLgSession clientSLgSession;
+  protected ServerSLgSession serverSLgSession;
 
   public void init(InputStream configStream, String clientID) throws Exception {
     try {
@@ -109,6 +112,11 @@ public abstract class AbstractImmediateClient extends TBase implements ClientSLg
     fail("Received \"PLA\" event, request[" + request + "], answer[" + answer + "], on session[" + session + "]", null);
   }
 
+  public void doLocationReportRequestEvent(ServerSLgSession session, LocationReportRequest request) throws InternalException, IllegalDiameterStateException,
+      RouteException, OverloadException {
+    fail("Received \"LRR\" event, request[" + request + "], on session[" + session + "]", null);
+  }
+
   // ----------- conf parts
 
   public String getSessionId() {
@@ -119,7 +127,15 @@ public abstract class AbstractImmediateClient extends TBase implements ClientSLg
     return this.clientSLgSession;
   }
 
-  // Attributes for Provide Location Request (PLR)
+  public void fetchSession(String sessionId) throws InternalException {
+    this.serverSLgSession = stack.getSession(sessionId, ServerSLgSession.class);
+  }
+
+  /*public ServerSLgSession getServerSLgSession() {
+    return this.serverSLgSession;
+  }*/
+
+  // Attributes for Provide Location Request (PLR) and Location Report Answer
   protected abstract int getSLgLocationType();
   protected abstract String getUserName(); // IE: IMSI
   protected abstract byte[] getMSISDN();
@@ -155,6 +171,7 @@ public abstract class AbstractImmediateClient extends TBase implements ClientSLg
   protected abstract int getPrioritizedListIndicator();
   protected abstract byte[] getVisitedPLMNId();
   protected abstract int getPeriodicLocationSupportIndicator();
+  protected abstract long getLRAFLags();
 
   // ----------- 3GPP TS 29.172 reference
 
@@ -442,5 +459,105 @@ public abstract class AbstractImmediateClient extends TBase implements ClientSLg
 
     return plr;
   }
+
+  // ----------- 3GPP TS 29.172 reference
+
+  public LocationReportAnswer createLRA(LocationReportRequest lrr, long resultCode) throws Exception {
+  /*
+  < Location-Report-Answer > ::=	< Diameter Header: 8388621, PXY, 16777255>
+
+    < Session-Id >
+	[ Vendor-Specific-Application-Id ]
+	[ Result-Code ]
+	[ Experimental-Result ]
+	{ Auth-Session-State }
+	{ Origin-Host }
+	{ Origin-Realm }
+	[ GMLC-Address ]
+	[ LRA-Flags ]
+	[ Reporting-PLMN-List ]
+	[ LCS-Reference-Number ]
+	*[ Supported-Features ]
+	*[ AVP ]
+	*[ Failed-AVP ]
+	*[ Proxy-Info ]
+	*[ Route-Record ]
+
+  */
+    LocationReportAnswer lra = new LocationReportAnswerImpl((Request) lrr.getMessage(), resultCode);
+
+    AvpSet reqSet = lrr.getMessage().getAvps();
+    AvpSet set = lra.getMessage().getAvps();
+    set.removeAvp(Avp.DESTINATION_HOST);
+    set.removeAvp(Avp.DESTINATION_REALM);
+    set.addAvp(reqSet.getAvp(Avp.AUTH_APPLICATION_ID));
+
+    // { Vendor-Specific-Application-Id }
+    if (set.getAvp(Avp.VENDOR_SPECIFIC_APPLICATION_ID) == null) {
+      AvpSet vendorSpecificApplicationId = set.addGroupedAvp(Avp.VENDOR_SPECIFIC_APPLICATION_ID, 0, false, false);
+      // 1* [ Vendor-Id ]
+      vendorSpecificApplicationId.addAvp(Avp.VENDOR_ID, getApplicationId().getVendorId(), true);
+      // 0*1{ Auth-Application-Id }
+      vendorSpecificApplicationId.addAvp(Avp.AUTH_APPLICATION_ID, getApplicationId().getAuthAppId(), true);
+    }
+    // [ Result-Code ]
+    // [ Experimental-Result ]
+    // { Auth-Session-State }
+    if (set.getAvp(Avp.AUTH_SESSION_STATE) == null) {
+      set.addAvp(Avp.AUTH_SESSION_STATE, 1);
+    }
+
+    //[ GMLC-Address ]
+    java.net.InetAddress gmlcAddress = getGMLCAddress();
+    if (gmlcAddress != null){
+      set.addAvp(Avp.GMLC_ADDRESS, gmlcAddress, 10415, false, false);
+    }
+
+    // [ LRA-Flags ]
+    long lraFlags = getLRAFLags();
+    if (lraFlags != -1){
+      set.addAvp(Avp.LRA_FLAGS, lraFlags, 10415, false, false, true);
+    }
+
+    //[ Reporting-PLMN-List ]
+/*
+  Reporting-PLMN-List ::= <AVP header: 2543 10415>
+    1*20{ PLMN-ID-List }
+    [ Prioritized-List-Indicator ]
+    *[ AVP ]
+
+  PLMN-ID-List ::= <AVP header: 2544 10415>
+    { Visited-PLMN-Id }
+    [ Periodic-Location-Support-Indicator ]
+    *[ AVP ]
+*/
+    AvpSet reportingPLMNList = set.addGroupedAvp(Avp.REPORTING_PLMN_LIST, 10415, false, false);
+    int prioritizedListIndicator = getPrioritizedListIndicator();
+    AvpSet plmnIdList = set.addGroupedAvp(Avp.PLMN_ID_LIST, 10415, false, false);
+    byte[] visitedPLMNId = getVisitedPLMNId();
+    int periodicLocationSupportIndicator = getPeriodicLocationSupportIndicator();
+
+    if (prioritizedListIndicator != -1){
+      reportingPLMNList.addAvp(Avp.PRIORITIZED_LIST_INDICATOR, prioritizedListIndicator, 10415, false, false);
+    }
+    if (plmnIdList != null){
+      reportingPLMNList.addAvp(plmnIdList);
+    }
+    if (visitedPLMNId != null){
+      plmnIdList.addAvp(Avp.VISITED_PLMN_ID, visitedPLMNId, 10415, false, false);
+    }
+    if (periodicLocationSupportIndicator != -1){
+      plmnIdList.addAvp(Avp.PERIODIC_LOCATION_SUPPORT_INDICATOR, periodicLocationSupportIndicator, 10415, false, false);
+    }
+
+    // [ LCS-Reference-Number ]
+    byte[] lcsReferenceNumber = getLCSReferenceNumber();
+    if (lcsReferenceNumber != null){
+      set.addAvp(Avp.LCS_REFERENCE_NUMBER, lcsReferenceNumber, 10415, true, false);
+    }
+
+    return lra;
+  }
+
 
 }
