+++ testsuite/tests/src/test/java/org/mobicents/diameter/stack/functional/slg/base/SLgSessionBasicPLRFlowTest.java
@@ -51,10 +51,10 @@ import org.junit.runners.Parameterized.Parameters;
  * @author <a href="mailto:baranowb@gmail.com"> Bartosz Baranowski </a>
  */
 @RunWith(Parameterized.class)
-public class SLgSessionBasicImmFlowTest {
+public class SLgSessionBasicPLRFlowTest {
   // TODO: add test on replicated nodes ?
-  private ClientImmediateSLg clientNode;
-  private ServerImmediateSLg serverNode1;
+  private ClientSLg clientNode;
+  private ServerSLg serverNode1;
   private URI clientConfigURI;
   private URI serverNode1ConfigURI;
 
@@ -64,7 +64,7 @@ public class SLgSessionBasicImmFlowTest {
    * // @param node2
    * // @param serverCount
    */
-  public SLgSessionBasicImmFlowTest(String clientConfigUrl, String serverNode1ConfigURL) throws Exception {
+  public SLgSessionBasicPLRFlowTest(String clientConfigUrl, String serverNode1ConfigURL) throws Exception {
     super();
     this.clientConfigURI = new URI(clientConfigUrl);
     this.serverNode1ConfigURI = new URI(serverNode1ConfigURL);
@@ -73,8 +73,8 @@ public class SLgSessionBasicImmFlowTest {
   @Before
   public void setUp() throws Exception {
     try {
-      this.clientNode = new ClientImmediateSLg();
-      this.serverNode1 = new ServerImmediateSLg();
+      this.clientNode = new ClientSLg();
+      this.serverNode1 = new ServerSLg();
 
       this.serverNode1.init(new FileInputStream(new File(this.serverNode1ConfigURI)), "SERVER1");
       this.serverNode1.start();
@@ -183,7 +183,7 @@ public class SLgSessionBasicImmFlowTest {
     //String replicatedClient = "configurations/functional-slh/replicated-config-client.xml";
     //String replicatedServer1 = "configurations/functional-slh/replicated-config-server-node1.xml";
 
-    Class<SLgSessionBasicImmFlowTest> t = SLgSessionBasicImmFlowTest.class;
+    Class<SLgSessionBasicPLRFlowTest> t = SLgSessionBasicPLRFlowTest.class;
     client = t.getClassLoader().getResource(client).toString();
     server1 = t.getClassLoader().getResource(server1).toString();
     //replicatedClient = t.getClassLoader().getResource(replicatedClient).toString();
